client: moved all chat methods to a separate service (`ChatService`)
