+++ client/src/app/services/auth.service.ts
@@ -3,13 +3,9 @@ import { BehaviorSubject, Observable } from 'rxjs';
 import {
   LoginRequest,
   RegisterRequest,
-  EditCellRequest,
-  SendGlobalMessageRequest,
-  GetMessageboxUsernamesRequest,
 } from '../../../../model/server-requests';
 import { User } from '../../../../model/user';
-import { Map, Cell } from '../../../../model/map';
-import { Message } from '../../../../model/message';
+import { Map } from '../../../../model/map';
 import { BackendService } from '../services/backend.service';
 import { map } from 'rxjs/operators';
 
@@ -69,48 +65,5 @@ export class AuthService {
   GetMap(): Observable<Map> {
     // if everything went okay, then the response contains map's data
     return this.backend.getMap().pipe(map((response) => response.map));
-  }
-
-  /**
-   * Edit cell data (owner, new building, level up, etc)
-   * Possible categories:
-   *    owner         (when buing/trading cells)
-   *    buildingType  (when buing building)
-   *    buildingLvl   (when upgrading building)
-   */
-  EditCell(
-    cellId: number,
-    changedCategory: string,
-    changedValue: number
-  ): Observable<never> {
-    const payload: EditCellRequest = {
-      cellId,
-      changedCategory,
-      changedValue,
-    };
-    return this.backend.editCell(payload);
-  }
-
-  SendGlobalMessage(username: string, message: string): Observable<never> {
-    const payload: SendGlobalMessageRequest = {
-      username,
-      message,
-    };
-    return this.backend.sendGlobalMessage(payload);
-  }
-
-  GetGlobalMessages(): Observable<Message[]> {
-    return this.backend
-      .getGlobalMessages()
-      .pipe(map((response) => response.messages));
-  }
-
-  GetUserChats(username: string): Observable<string[]> {
-    const payload: GetMessageboxUsernamesRequest = {
-      username,
-    };
-    return this.backend
-      .getUserChats(payload)
-      .pipe(map((response) => response.usernames));
   }
 }
