+++ client/src/app/services/backend.service.ts
@@ -1,7 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Message } from '../../../../model/message';
 import {
   LoginRequest,
   RegisterRequest,
